+++ src/Whoami/Service/AnyLib.hs
@@ -10,7 +10,7 @@ import           Whoami.Service.Data.Info       (Library (..), ServiceType)
 import           Whoami.Service.Internal.Fetch  (fetchHtml)
 import           Whoami.Service.Internal.Scrape (scrapeDesc)
 
-newtype AnyLib = AnyLib LibConfig
+newtype AnyLib = AnyLib LibConfig deriving (Show)
 
 libs :: Proxy AnyLib
 libs = Proxy