+++ src/Whoami/Service/Internal/Utils.hs
@@ -5,7 +5,6 @@ module Whoami.Service.Internal.Utils where
 import           RIO
 
 import           Data.Extensible
-import qualified Shelly          as S
 
 embedM :: (Functor f, x ∈ xs) => Comp f h x -> f (h :| xs)
 embedM = fmap embed . getComp
@@ -17,4 +16,4 @@ valid :: (a -> Bool) -> a -> Maybe a
 valid p a = if p a then pure a else Nothing
 
 sleep :: MonadIO m => Int -> m ()
-sleep = S.shelly . S.sleep
+sleep = liftIO . threadDelay . (* 1_000_000)