+++ src/Whoami/Service/Qiita.hs
@@ -36,14 +36,13 @@ fetchQiitaPosts = do
   account <- Map.lookup "qiita" <$> asks (view #account . view #config)
   case account of
     Just name -> fetchQiitaPosts' name
-    Nothing   -> throwIO $ ServiceException "qiita account is not defined"
+    Nothing   -> throwM $ ServiceException "qiita account is not defined"
 
 fetchQiitaPosts' :: Text -> ServiceM [QiitaPost]
 fetchQiitaPosts' name = do
-  num <- fromMaybe 100 <$> asks (view #count . view #qiita . view #config)
-  let
-    url = https "qiita.com" /: "api" /: "v2" /: "users" /: name /: "items"
-    params = "per_page" =: num
+  num <- fromMaybe 10 <$> asks (view #count . view #qiita . view #config)
+  let url = https "qiita.com" /: "api" /: "v2" /: "users" /: name /: "items"
+      params = "per_page" =: num
   result <- runReq' defaultHttpConfig $ req GET url NoReqBody jsonResponse params
   case result of
     Left err   -> throwFetchError (Left err)