+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+- Feat: collect medium posts by user id
+- Refactor: remove dependent for shelly using threadDelay
+- Refactor: use throwM instead of throwIO
+
 ## 0.2.0.0
 
 - use mix library